demo/backend/voxtral-asr/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
numpy
mistralai[realtime]
python-dotenv

